+++ src/app/[userNickname]/[listId]/_components/ListDetailOuter/DeleteModalButton.css.ts
